backend/requirements.txt
# backend/requirements.txt

# FastAPI and server (Compatible versions)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# PDF Processing
PyPDF2==3.0.1

# AI Integration
google-generativeai==0.3.0

# HTTP Clients (for fetching content)
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data validation (Compatible with FastAPI)
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic-core==2.14.6

# Utilities
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
reportlab==4.0.7  # For creating test PDFs

# Optional: If you want to use BeautifulSoup for web scraping later
beautifulsoup4==4.12.2
lxml==4.9.3